requirements.txt
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-ai>=0.1.0  # This is a hypothetical package name, you might need to adjust based on actual package name
